Split mobile browser pane modules
