+++ config/max-lines-baseline.txt
@@ -83,7 +83,6 @@ mobile-config app/h/*/tasks.tsx
 mobile-config app/index.tsx
 mobile-config app/troubleshoot.tsx
 mobile-config scripts/mock-server.ts
-mobile-config src/browser/MobileBrowserPane.tsx
 mobile-config src/components/NewWorktreeModal.tsx
 mobile-config src/terminal/terminal-webview-html.ts
 mobile-config src/transport/rpc-client.ts
