+++ mobile/.oxlintrc.json
@@ -52,12 +52,6 @@
       }
     },
     {
-      "files": ["src/browser/MobileBrowserPane.tsx"],
-      "rules": {
-        "max-lines": ["error", { "max": 1594, "skipBlankLines": true, "skipComments": true }]
-      }
-    },
-    {
       "files": ["app/index.tsx"],
       "rules": {
         "max-lines": ["error", { "max": 1422, "skipBlankLines": true, "skipComments": true }]
